+++ src/lang/en.lua
@@ -3,127 +3,128 @@
 
 AddLocale('en', {
     HeaderText = " DPS / Total Damage: ",
-    AttackText = "Attack",
-    DashStrikeText = "Dash-Strike",
-    SpecialText = "Special",
-    CastText = "Cast",
-    DashText = "Dash",
-    WallSlam = "Wall Slam",
     EnemyText = "Enemy",
-    Charm = "Charm",
-    -- npc assists
-    ArtemisName = "Artemis",
-    NemesisName = "Nemesis",
-    HeraclesName = "Heracles",
-    IcarusName = "Icarus",
-    -- familiars
-    Frinos = "Frinos",
-    Toula = "Toula",
-    -- traps
-    SteamTrap = "Steam Trap",
-    OilFire = "Oil Fire",
-    FireBarrel = "Fire Barrel",
-    SpikeTrap = "Spike Trap",
-    BrambleTrap = "Brambles",
-    ThornTrap = "Thorns",
-    Splinters = "Splinters",
-    BeamTrap = "Beam",
-    Pillar = "Collapsing Pillar",
-    Lava = "Lava",
-    Rubble = "Falling Rubble",
-    GenericTrap = "Trap",
-    --  misc
+    WallSlam = "Wall Slam",
+    EphyraPylon = "Pylon Spirits",
     SheepAttack = "Sheep Attack",
     Boulder = "Polyphemus Boulder",
-    ShadeSprint = "Necromantic Influence",
-    EphyraPylon = "Pylon Spirits",
-    -- these correspond with Id/DisplayName in TraitText.[lang].sjson or HelpText.[lang].sjson
-    -- aphrodite
-    AphroditeWeaponBoon = "Flutter Strike",
-    AphroditeSpecialBoon = "Flutter Flourish",
-    AphroditeCastBoon = "Rapture Ring",
-    AphroditeSprintBoon = "Passion Dash",
-    ManaBurstBoon = "Heart Breaker",
-    -- apollo
-    ApolloWeaponBoon = "Nova Strike",
-    ApolloSpecialBoon = "Nova Flourish",
-    ApolloCastBoon = "Solar Ring",
-    ApolloSprintBoon = "Blinding Sprint",
-    ApolloRetaliateBoon = "Light Smite",
-    ApolloMissStrikeBoon = "Critical Miss",
-    -- artemis
-    SupportingFireBoon = "Support Fire",
-    OmegaCastVolleyBoon = "Easy Shot",
-    -- demeter
-    DemeterWeaponBoon = "Ice Strike",
-    DemeterSpecialBoon = "Ice Flourish",
-    DemeterCastBoon = "Arctic Ring",
-    DemeterSprintBoon = "Frigid Sprint",
-    InstantRootKill = "Winter Harvest",
-    -- hades
-    HadesCastProjectileBoon = "Howling Soul",
-    HadesPreDamageBoon = "Old Grudge",
-    -- hera
-    HeraWeaponBoon = "Sworn Strike",
-    HeraSpecialBoon = "Sworn Flourish",
-    HeraCastBoon = "Engagement Ring",
-    HeraSprintBoon = "Nexus Sprint", -- not currently used
-    DamageShareRetaliateBoon = "Nasty Comeback",
-    Link = "Hitch", --HelpText
-    LinkedDeathDamageBoon = "Dying Wish",
-    -- hestia
-    HestiaWeaponBoon = "Flame Strike",
-    HestiaSpecialBoon = "Flame Flourish",
-    HestiaSprintBoon = "Soot Sprint",
-    HestiaCastBoon = "Smolder Ring",
-    BurnExplodeBoon = "Natural Gas",
-    Burn = "Scorch", -- HelpText
-    FireballManaSpecialBoon = "Controlled Burn",
-    CastProjectileBoon = "Glowing Coal",
-    BurnConsumeBoon = "Fire Extinguisher",
-    -- hephaestus
-    HephaestusWeaponBoon = "Volcanic Strike",
-    HephaestusSpecialBoon = "Volcanic Flourish",
-    HephaestusCastBoon = "Anvil Ring",
-    HephaestusSprintBoon = "Smithy Sprint",
-    MassiveSlam_Name = "Blast",
-    MassiveKnockupBoon = "Furnace Blast",
-    -- icarus
-    OmegaExplodeBoon = "Explosive Intent",
-    CastHazardBoon = "Hazard Boom",
-    BreakExplosiveArmorBoon = "Destructive Coating",
-    -- medea
-    DeathDefianceRetaliateCurse = "Malice in Kind",
-    SpawnDamageCurse = "Suffering on Sight",
-    ArmorPenaltyCurse = "Corrosion on Sight",
-    -- odysseus
-    BossPreDamageKeepsake = "Knuckle Bones",
-    -- poseidon
-    PoseidonWeaponBoon = "Wave Strike",
-    PoseidonSpecialBoon = "Wave Flourish",
-    PoseidonCastBoon = "Geyser Ring",
-    PoseidonSprintBoon = "Breaker Sprint",
-    PoseidonSplash_Name = "Waves",
-    SlamExplosionBoon = "Crashing Wave",
-    -- selene
-    SpellLaserTrait = "Lunar Ray",
-    SpellLeapTrait = "Wolf Howl",
-    SpellMeteorTrait = "Total Eclipse",
-    SpellTransformTrait = "Dark Side",
-    -- zeus
-    ZeusWeaponBoon = "Heaven Strike",
-    ZeusSpecialBoon = "Heaven Flourish",
-    ZeusSprintBoon = "Thunder Sprint",
-    ZeusCastBoon = "Storm Ring",
-    Echo = "Blitz", -- HelpText
-    FocusLightningBoon = "Static Shock",
-    ZeusManaBoltBoon = "Spirit Surge",
-    BoltRetaliateBoon = "Divine Vengeance",
-    CastAnywhereBoon = "Lightning Lance",
-    SpawnKillBoon = "Shocking Loss",
-    -- duo
-    ShadeMercFireballBoon = "Burning Desire", -- aphrodite x hestia
-    PoseidonSplashSprintBoon = "Beach Ball",  -- apollo x poseidon
-    MassiveCastBoon = "Seismic Hammer",  -- hephaestus x poseidon
-    SteamBoon = "Scalding Vapor" -- hestia x poseidon
+    -- everything under this line is no longer used!
+    -- AttackText = "Attack",
+    -- DashStrikeText = "Dash-Strike",
+    -- SpecialText = "Special",
+    -- CastText = "Cast",
+    -- DashText = "Dash",
+    -- Charm = "Charm",
+    -- -- npc assists
+    -- ArtemisName = "Artemis",
+    -- NemesisName = "Nemesis",
+    -- HeraclesName = "Heracles",
+    -- IcarusName = "Icarus",
+    -- -- familiars
+    -- Frinos = "Frinos",
+    -- Toula = "Toula",
+    -- -- traps
+    -- SteamTrap = "Steam Trap",
+    -- OilFire = "Oil Fire",
+    -- FireBarrel = "Fire Barrel",
+    -- SpikeTrap = "Spike Trap",
+    -- BrambleTrap = "Brambles",
+    -- ThornTrap = "Thorns",
+    -- Splinters = "Splinters",
+    -- BeamTrap = "Beam",
+    -- Pillar = "Collapsing Pillar",
+    -- Lava = "Lava",
+    -- Rubble = "Falling Rubble",
+    -- GenericTrap = "Trap",
+    -- --  misc
+    -- ShadeSprint = "Necromantic Influence",
+    -- -- these correspond with Id/DisplayName in TraitText.[lang].sjson or HelpText.[lang].sjson
+    -- -- aphrodite
+    -- AphroditeWeaponBoon = "Flutter Strike",
+    -- AphroditeSpecialBoon = "Flutter Flourish",
+    -- AphroditeCastBoon = "Rapture Ring",
+    -- AphroditeSprintBoon = "Passion Dash",
+    -- ManaBurstBoon = "Heart Breaker",
+    -- -- apollo
+    -- ApolloWeaponBoon = "Nova Strike",
+    -- ApolloSpecialBoon = "Nova Flourish",
+    -- ApolloCastBoon = "Solar Ring",
+    -- ApolloSprintBoon = "Blinding Sprint",
+    -- ApolloRetaliateBoon = "Light Smite",
+    -- ApolloMissStrikeBoon = "Critical Miss",
+    -- -- artemis
+    -- SupportingFireBoon = "Support Fire",
+    -- OmegaCastVolleyBoon = "Easy Shot",
+    -- -- demeter
+    -- DemeterWeaponBoon = "Ice Strike",
+    -- DemeterSpecialBoon = "Ice Flourish",
+    -- DemeterCastBoon = "Arctic Ring",
+    -- DemeterSprintBoon = "Frigid Sprint",
+    -- InstantRootKill = "Winter Harvest",
+    -- -- hades
+    -- HadesCastProjectileBoon = "Howling Soul",
+    -- HadesPreDamageBoon = "Old Grudge",
+    -- -- hera
+    -- HeraWeaponBoon = "Sworn Strike",
+    -- HeraSpecialBoon = "Sworn Flourish",
+    -- HeraCastBoon = "Engagement Ring",
+    -- HeraSprintBoon = "Nexus Sprint", -- not currently used
+    -- DamageShareRetaliateBoon = "Nasty Comeback",
+    -- Link = "Hitch", --HelpText
+    -- LinkedDeathDamageBoon = "Dying Wish",
+    -- -- hestia
+    -- HestiaWeaponBoon = "Flame Strike",
+    -- HestiaSpecialBoon = "Flame Flourish",
+    -- HestiaSprintBoon = "Soot Sprint",
+    -- HestiaCastBoon = "Smolder Ring",
+    -- BurnExplodeBoon = "Natural Gas",
+    -- Burn = "Scorch", -- HelpText
+    -- FireballManaSpecialBoon = "Controlled Burn",
+    -- CastProjectileBoon = "Glowing Coal",
+    -- BurnConsumeBoon = "Fire Extinguisher",
+    -- -- hephaestus
+    -- HephaestusWeaponBoon = "Volcanic Strike",
+    -- HephaestusSpecialBoon = "Volcanic Flourish",
+    -- HephaestusCastBoon = "Anvil Ring",
+    -- HephaestusSprintBoon = "Smithy Sprint",
+    -- MassiveSlam_Name = "Blast",
+    -- MassiveKnockupBoon = "Furnace Blast",
+    -- -- icarus
+    -- OmegaExplodeBoon = "Explosive Intent",
+    -- CastHazardBoon = "Hazard Boom",
+    -- BreakExplosiveArmorBoon = "Destructive Coating",
+    -- -- medea
+    -- DeathDefianceRetaliateCurse = "Malice in Kind",
+    -- SpawnDamageCurse = "Suffering on Sight",
+    -- ArmorPenaltyCurse = "Corrosion on Sight",
+    -- -- odysseus
+    -- BossPreDamageKeepsake = "Knuckle Bones",
+    -- -- poseidon
+    -- PoseidonWeaponBoon = "Wave Strike",
+    -- PoseidonSpecialBoon = "Wave Flourish",
+    -- PoseidonCastBoon = "Geyser Ring",
+    -- PoseidonSprintBoon = "Breaker Sprint",
+    -- PoseidonSplash_Name = "Waves",
+    -- SlamExplosionBoon = "Crashing Wave",
+    -- -- selene
+    -- SpellLaserTrait = "Lunar Ray",
+    -- SpellLeapTrait = "Wolf Howl",
+    -- SpellMeteorTrait = "Total Eclipse",
+    -- SpellTransformTrait = "Dark Side",
+    -- -- zeus
+    -- ZeusWeaponBoon = "Heaven Strike",
+    -- ZeusSpecialBoon = "Heaven Flourish",
+    -- ZeusSprintBoon = "Thunder Sprint",
+    -- ZeusCastBoon = "Storm Ring",
+    -- Echo = "Blitz", -- HelpText
+    -- FocusLightningBoon = "Static Shock",
+    -- ZeusManaBoltBoon = "Spirit Surge",
+    -- BoltRetaliateBoon = "Divine Vengeance",
+    -- CastAnywhereBoon = "Lightning Lance",
+    -- SpawnKillBoon = "Shocking Loss",
+    -- -- duo
+    -- ShadeMercFireballBoon = "Burning Desire", -- aphrodite x hestia
+    -- PoseidonSplashSprintBoon = "Beach Ball",  -- apollo x poseidon
+    -- MassiveCastBoon = "Seismic Hammer",  -- hephaestus x poseidon
+    -- SteamBoon = "Scalding Vapor" -- hestia x poseidon
 })