+++ src/lang/fr.lua
@@ -3,127 +3,128 @@
 
 AddLocale('fr', {
     HeaderText = " DPS / Dégâts totaux : ",
-    AttackText = "Attaques",
-    DashStrikeText = "Assaut",
-    SpecialText = "Techniques",
-    CastText = "Glyphes",
-    DashText = "Ruée",
-    WallSlam = "Choc Contre Les Murs",
     EnemyText = "Ennemis",
-    Charm = "Subjugué",
-    -- npc assists
-    ArtemisName = "Artémis",
-    NemesisName = "Némésis",
-    HeraclesName = "Héraclès",
-    IcarusName = "Icarus",
-    -- familiars
-    Frinos = "Frinos",
-    Toula = "Toula",
-    -- traps
-    SteamTrap = "Vapeurs",
-    OilFire = "Feu de flaque d'huile",
-    FireBarrel = "Tonneaux de feu",
-    SpikeTrap = "Pics",
-    BrambleTrap = "Ronces",
-    ThornTrap = "Épines",
-    Splinters = "Éclats d'Arbres",
-    BeamTrap = "Poutres",
-    Pillar = "Piliers effondrés",
-    Lava = "Lave",
-    Rubble = "Chute de gravats",
-    GenericTrap = "Pièges",
-    --  misc
+    WallSlam = "Choc Contre Les Murs",
+    EphyraPylon = "Pylône Spirituel",
     SheepAttack = "Dégâts laineux",
     Boulder = "Polyphème",
-    ShadeSprint = "Influence Nécromantique",
-    EphyraPylon = "Pylône Spirituel",
-    -- these correspond with Id/DisplayName in TraitText.[lang].sjson or HelpText.[lang].sjson
-    -- aphrodite
-    AphroditeWeaponBoon = "Frappe Séduisante",
-    AphroditeSpecialBoon = "Technique Séduisante",
-    AphroditeCastBoon = "Glyphe Attirant",
-    AphroditeSprintBoon = "Élan Fougueux",
-    ManaBurstBoon = "Passion Dévorante",
-    -- apollo
-    ApolloWeaponBoon = "Frappe Éclatante",
-    ApolloSpecialBoon = "Technique Éclatante",
-    ApolloCastBoon = "Glyphe Solaire",
-    ApolloSprintBoon = "Ruée Aveuglante",
-    ApolloRetaliateBoon = "Coup De Soleil",
-    ApolloMissStrikeBoon = "Échec critique",
-    -- artemis
-    SupportingFireBoon = "Tir De Couverture",
-    OmegaCastVolleyBoon = "Cible Facile",
-    -- demeter
-    DemeterWeaponBoon = "Frappe Glaciale",
-    DemeterSpecialBoon = "Technique Glaciale",
-    DemeterCastBoon = "Glyphe Polaire",
-    DemeterSprintBoon = "Élant Frigorifiant",
-    InstantRootKill = "Dépression Soudaine",
-    -- hades
-    HadesCastProjectileBoon = "Âme Rugissante",
-    HadesPreDamageBoon = "Haine Ancétrale",
-    -- hera
-    HeraWeaponBoon = "Frappe Unificatrice",
-    HeraSpecialBoon = "Technique Unificatrice",
-    HeraCastBoon = "Glyphe Cérémoniel",
-    HeraSprintBoon = "Ruée Cohésive", -- not currently used
-    DamageShareRetaliateBoon = "Retour en force",
-    Link = "Malheur Héréditaire", --HelpText
-    LinkedDeathDamageBoon = "Dernière Volonté",
-    -- hestia
-    HestiaWeaponBoon = "Frappe Emflammée",
-    HestiaSpecialBoon = "Technique Emflammée",
-    HestiaSprintBoon = "Ruée Calcinante",
-    HestiaCastBoon = "Glyphe Inférnale",
-    BurnExplodeBoon = "Gaz Naturel",
-    Burn = "Brûlure", -- HelpText
-    FireballManaSpecialBoon = "Brûlage Dirigé",
-    CastProjectileBoon = "Charbon Ardent",
-    BurnConsumeBoon = "Extinction Brutale",
-    -- hephaestus
-    HephaestusWeaponBoon = "Frappe Volcanique",
-    HephaestusSpecialBoon = "Technique Volcanique",
-    HephaestusCastBoon = "Glyphe Martelant",
-    HephaestusSprintBoon = "Ruée Assourdissante",
-    MassiveSlam_Name = "Explosion",
-    MassiveKnockupBoon = "Décharge Thermique",
-    -- icarus
-    OmegaExplodeBoon = "Humeur Explosive",
-    CastHazardBoon = "Dangereuse Déflagration",
-    BreakExplosiveArmorBoon = "Revêtement Destructeur",
-    -- medea
-    DeathDefianceRetaliateCurse = "Œil pour Œil",
-    SpawnDamageCurse = "Souffrance Subite",
-    ArmorPenaltyCurse = "Corrosion Subite",
-    -- odysseus
-    BossPreDamageKeepsake = "Osselets",
-    -- poseidon
-    PoseidonWeaponBoon = "Frappe Marine",
-    PoseidonSpecialBoon = "Technique Marine",
-    PoseidonCastBoon = "Glyphe Bouillonant",
-    PoseidonSprintBoon = "Ruée Torrentiel",
-    PoseidonSplash_Name = "Vagues de Poséidon",
-    SlamExplosionBoon = "Vague Déferlante",
-    -- selene
-    SpellLaserTrait = "Rayon lunaire",
-    SpellLeapTrait = "Hulement Bestial",
-    SpellMeteorTrait = "Éclipse Totale",
-    SpellTransformTrait = "Face Cachée",
-    -- zeus
-    ZeusWeaponBoon = "Frappe Céleste",
-    ZeusSpecialBoon = "Technique Céleste",
-    ZeusSprintBoon = "Ruée Foudroyante",
-    ZeusCastBoon = "Glyphe Orageux",
-    Echo = "Chargé", -- HelpText
-    FocusLightningBoon = "Électricité Statique",
-    ZeusManaBoltBoon = "Choc Psychique",
-    BoltRetaliateBoon = "Vengeance Divine",
-    CastAnywhereBoon = "Lance Fulminante",
-    SpawnKillBoon = "Châtiment Tonitruant",
-    -- duo
-    ShadeMercFireballBoon = "Ambiance Torride", -- aphrodite x hestia
-    PoseidonSplashSprintBoon = "Ballet Aquatique",  -- apollo x poseidon
-    MassiveCastBoon = "Marteau Sismique",  -- hephaestus x poseidon
-    SteamBoon = "Cuisson Vapeur" -- hestia x poseidon
+    -- everything under this line is no longer used!
+    -- AttackText = "Attaques",
+    -- DashStrikeText = "Assaut",
+    -- SpecialText = "Techniques",
+    -- CastText = "Glyphes",
+    -- DashText = "Ruée",
+    -- Charm = "Subjugué",
+    -- -- npc assists
+    -- ArtemisName = "Artémis",
+    -- NemesisName = "Némésis",
+    -- HeraclesName = "Héraclès",
+    -- IcarusName = "Icarus",
+    -- -- familiars
+    -- Frinos = "Frinos",
+    -- Toula = "Toula",
+    -- -- traps
+    -- SteamTrap = "Vapeurs",
+    -- OilFire = "Feu de flaque d'huile",
+    -- FireBarrel = "Tonneaux de feu",
+    -- SpikeTrap = "Pics",
+    -- BrambleTrap = "Ronces",
+    -- ThornTrap = "Épines",
+    -- Splinters = "Éclats d'Arbres",
+    -- BeamTrap = "Poutres",
+    -- Pillar = "Piliers effondrés",
+    -- Lava = "Lave",
+    -- Rubble = "Chute de gravats",
+    -- GenericTrap = "Pièges",
+    -- --  misc
+    -- ShadeSprint = "Influence Nécromantique",
+    -- -- these correspond with Id/DisplayName in TraitText.[lang].sjson or HelpText.[lang].sjson
+    -- -- aphrodite
+    -- AphroditeWeaponBoon = "Frappe Séduisante",
+    -- AphroditeSpecialBoon = "Technique Séduisante",
+    -- AphroditeCastBoon = "Glyphe Attirant",
+    -- AphroditeSprintBoon = "Élan Fougueux",
+    -- ManaBurstBoon = "Passion Dévorante",
+    -- -- apollo
+    -- ApolloWeaponBoon = "Frappe Éclatante",
+    -- ApolloSpecialBoon = "Technique Éclatante",
+    -- ApolloCastBoon = "Glyphe Solaire",
+    -- ApolloSprintBoon = "Ruée Aveuglante",
+    -- ApolloRetaliateBoon = "Coup De Soleil",
+    -- ApolloMissStrikeBoon = "Échec critique",
+    -- -- artemis
+    -- SupportingFireBoon = "Tir De Couverture",
+    -- OmegaCastVolleyBoon = "Cible Facile",
+    -- -- demeter
+    -- DemeterWeaponBoon = "Frappe Glaciale",
+    -- DemeterSpecialBoon = "Technique Glaciale",
+    -- DemeterCastBoon = "Glyphe Polaire",
+    -- DemeterSprintBoon = "Élant Frigorifiant",
+    -- InstantRootKill = "Dépression Soudaine",
+    -- -- hades
+    -- HadesCastProjectileBoon = "Âme Rugissante",
+    -- HadesPreDamageBoon = "Haine Ancétrale",
+    -- -- hera
+    -- HeraWeaponBoon = "Frappe Unificatrice",
+    -- HeraSpecialBoon = "Technique Unificatrice",
+    -- HeraCastBoon = "Glyphe Cérémoniel",
+    -- HeraSprintBoon = "Ruée Cohésive", -- not currently used
+    -- DamageShareRetaliateBoon = "Retour en force",
+    -- Link = "Malheur Héréditaire", --HelpText
+    -- LinkedDeathDamageBoon = "Dernière Volonté",
+    -- -- hestia
+    -- HestiaWeaponBoon = "Frappe Emflammée",
+    -- HestiaSpecialBoon = "Technique Emflammée",
+    -- HestiaSprintBoon = "Ruée Calcinante",
+    -- HestiaCastBoon = "Glyphe Inférnale",
+    -- BurnExplodeBoon = "Gaz Naturel",
+    -- Burn = "Brûlure", -- HelpText
+    -- FireballManaSpecialBoon = "Brûlage Dirigé",
+    -- CastProjectileBoon = "Charbon Ardent",
+    -- BurnConsumeBoon = "Extinction Brutale",
+    -- -- hephaestus
+    -- HephaestusWeaponBoon = "Frappe Volcanique",
+    -- HephaestusSpecialBoon = "Technique Volcanique",
+    -- HephaestusCastBoon = "Glyphe Martelant",
+    -- HephaestusSprintBoon = "Ruée Assourdissante",
+    -- MassiveSlam_Name = "Explosion",
+    -- MassiveKnockupBoon = "Décharge Thermique",
+    -- -- icarus
+    -- OmegaExplodeBoon = "Humeur Explosive",
+    -- CastHazardBoon = "Dangereuse Déflagration",
+    -- BreakExplosiveArmorBoon = "Revêtement Destructeur",
+    -- -- medea
+    -- DeathDefianceRetaliateCurse = "Œil pour Œil",
+    -- SpawnDamageCurse = "Souffrance Subite",
+    -- ArmorPenaltyCurse = "Corrosion Subite",
+    -- -- odysseus
+    -- BossPreDamageKeepsake = "Osselets",
+    -- -- poseidon
+    -- PoseidonWeaponBoon = "Frappe Marine",
+    -- PoseidonSpecialBoon = "Technique Marine",
+    -- PoseidonCastBoon = "Glyphe Bouillonant",
+    -- PoseidonSprintBoon = "Ruée Torrentiel",
+    -- PoseidonSplash_Name = "Vagues de Poséidon",
+    -- SlamExplosionBoon = "Vague Déferlante",
+    -- -- selene
+    -- SpellLaserTrait = "Rayon lunaire",
+    -- SpellLeapTrait = "Hulement Bestial",
+    -- SpellMeteorTrait = "Éclipse Totale",
+    -- SpellTransformTrait = "Face Cachée",
+    -- -- zeus
+    -- ZeusWeaponBoon = "Frappe Céleste",
+    -- ZeusSpecialBoon = "Technique Céleste",
+    -- ZeusSprintBoon = "Ruée Foudroyante",
+    -- ZeusCastBoon = "Glyphe Orageux",
+    -- Echo = "Chargé", -- HelpText
+    -- FocusLightningBoon = "Électricité Statique",
+    -- ZeusManaBoltBoon = "Choc Psychique",
+    -- BoltRetaliateBoon = "Vengeance Divine",
+    -- CastAnywhereBoon = "Lance Fulminante",
+    -- SpawnKillBoon = "Châtiment Tonitruant",
+    -- -- duo
+    -- ShadeMercFireballBoon = "Ambiance Torride", -- aphrodite x hestia
+    -- PoseidonSplashSprintBoon = "Ballet Aquatique",  -- apollo x poseidon
+    -- MassiveCastBoon = "Marteau Sismique",  -- hephaestus x poseidon
+    -- SteamBoon = "Cuisson Vapeur" -- hestia x poseidon
 })